Renamed namespace Static to Facade. Updated Facade\Engine, and tests.
