+++ phpunit_coverage.txt
@@ -1,12 +1,12 @@
 
 
 Code Coverage Report:
-  2025-05-08 09:18:33
+  2025-05-27 08:54:43
 
  Summary:
-  Classes: 68.63% (35/51)
-  Methods: 72.79% (222/305)
-  Lines:   68.83% (870/1264)
+  Classes: 66.67% (34/51)
+  Methods: 68.27% (213/312)
+  Lines:   65.71% (849/1292)
 
 Projom\Storage\Engine\Driver\Config
   Methods: 100.00% ( 4/ 4)   Lines: 100.00% (  9/  9)
@@ -24,10 +24,12 @@ Projom\Storage\Engine\Driver\DriverFactory
   Methods:  75.00% ( 3/ 4)   Lines:  84.21% ( 16/ 19)
 Projom\Storage\Engine\Driver\MySQL
   Methods:  87.50% (14/16)   Lines:  97.96% ( 96/ 98)
+Projom\Storage\Facade\Engine
+  Methods: 100.00% ( 4/ 4)   Lines: 100.00% (  8/  8)
 Projom\Storage\Logger\SimpleLogger
   Methods:  61.54% ( 8/13)   Lines:  79.37% ( 50/ 63)
 Projom\Storage\MySQL\Util
-  Methods: 100.00% ( 2/ 2)   Lines: 100.00% (  4/  4)
+  Methods: 100.00% ( 5/ 5)   Lines: 100.00% ( 13/ 13)
 Projom\Storage\Query\Util
   Methods: 100.00% ( 1/ 1)   Lines: 100.00% (  3/  3)
 Projom\Storage\SQL\Component\Column
@@ -86,11 +88,5 @@ Projom\Storage\SQL\Util\Operator
   Methods: 100.00% ( 1/ 1)   Lines: 100.00% (  1/  1)
 Projom\Storage\SQL\Util\Sort
   Methods: 100.00% ( 1/ 1)   Lines: 100.00% (  1/  1)
-Projom\Storage\Static\Engine
-  Methods:  87.50% ( 7/ 8)   Lines:  96.43% ( 27/ 28)
-Projom\Storage\Static\MySQL\Query
-  Methods: 100.00% ( 6/ 6)   Lines: 100.00% (  7/  7)
-Projom\Storage\Static\Query\DB
-  Methods: 100.00% ( 3/ 3)   Lines: 100.00% (  4/  4)
 Projom\Storage\Util
   Methods: 100.00% (12/12)   Lines: 100.00% ( 31/ 31)
